fm-webui-pytest/requirements.txt
pytest~=7.2.2
requests~=2.28.2
PyYAML~=6.0
dataclasses-json~=0.6.0
assertpy~=1.1